GCYS and GCYM dependency removal

+++ src/main/java/tekcays_addon/api/utils/roasting/RoastableMaterial.java
@@ -2,7 +2,6 @@ package tekcays_addon.api.utils.roasting;
 
 import gregtech.api.unification.material.Material;
 
-import static gregicality.science.api.unification.materials.GCYSMaterials.*;
 import static gregtech.api.unification.material.Materials.*;
 import static tekcays_addon.api.unification.TKCYAMaterials.*;
 import static tekcays_addon.api.utils.TKCYAValues.ATMOSPHERIC_PRESSURE;
@@ -30,24 +29,24 @@ public class RoastableMaterial {
 
     private final Material material;
     private final int temperature;
-    private final double pressure;
+    private final long pressure;
     private Material output;
     private Material fluidOutput;
 
-    public RoastableMaterial(Material material, int temperature, double pressure) {
+    public RoastableMaterial(Material material, int temperature, long pressure) {
         this.material = material;
         this.temperature = temperature;
         this.pressure = pressure;
     }
 
-    public RoastableMaterial(Material material, int temperature, double pressure, Material output) {
+    public RoastableMaterial(Material material, int temperature, long pressure, Material output) {
         this.material = material;
         this.temperature = temperature;
         this.pressure = pressure;
         this.output = output;
     }
 
-    public RoastableMaterial(Material material, int temperature, double pressure, Material output, Material fluidOutput) {
+    public RoastableMaterial(Material material, int temperature, long pressure, Material output, Material fluidOutput) {
         this.material = material;
         this.temperature = temperature;
         this.pressure = pressure;
@@ -63,7 +62,7 @@ public class RoastableMaterial {
         return this.temperature;
     }
 
-    public double getRoastingPressure() {
+    public long getRoastingPressure() {
         return this.pressure;
     }
     public Material getOutput() {
